Bug 61190 - [Bidi] Shortcut keys doesn't work with Arabic and other non-English keyboard
Backport fixes into 4.7.2.

+++ bundles/org.eclipse.swt/Eclipse SWT/gtk/org/eclipse/swt/widgets/Control.java
@@ -5889,7 +5889,7 @@ boolean translateMnemonic (Event event, Control control) {
 }
 
 boolean translateMnemonic (int keyval, GdkEventKey gdkEvent) {
-	int key = OS.gdk_keyval_to_unicode (keyval);
+	long key = OS.gdk_keyval_to_unicode (keyval);
 	if (key < 0x20) return false;
 	if (gdkEvent.state == 0) {
 		int code = traversalCode (keyval, gdkEvent);
